+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Security.
 
 ## [Unreleased]
 
+### Added
+
+- Added the Alchemist character model to combat, replacing the Ironclad placeholder. It idles, blinks and reacts when hit
+- Added the Alchemist character model to the Rest Site and the Merchant, replacing the Ironclad placeholders
+
 ## [0.6.1] - 2026-08-09
 
 ### Added